Added host script to run "vite --host" to package.json
Added a new npm run command to be able to test the website on real devices instead of only the dev tool. Was having an issue with the dev tool pixel 5 showing that users could scroll horizontally when they shouldn't have been able to. This is correct on the real phone so it is not an actual issue.