Updated CartContext and NavigationBar with new NavigationCart component
CartContext will now check local storage for an instance of shoppingCart and initialize the shoppingCart if on is found.

Added new NavigationCart component that will update with count of items in shoppingCart.

Updated NavigationBar to replace the user account icon with the NavigationCart when there are items in the cart.