Updated prebuiltCheckout to no longer remove order_id's
The order_id value was not being stored in the order document, only the user order's array. When the order was updated, the update to the user was erasing the order_id field and causing issues client side. This change preserves the older order_id on the user Orders array.